feat✨: add and deploy flag "split_days"
This flag is used to decide if input sequence of multiple days is split into 2D-tensor aligned along time of day of each data point.
This tensor will be fed into a Conv2D layer and generate a 1D-tensor with length of the number of samples in a day which is then fed to the given model.